backend/requirements.txt
fastapi
uvicorn
pydantic
swisseph
pyswisseph
timezonefinder
pytz
requests
google-genai
python-dotenv
